requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
openai==0.27.0
python-jose==3.3.0
cryptography==3.4.7
